# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Business API",
    "version": "2.0.0",
    "description": "Enterprise-grade middleware API for CAC.gov.ng business registration. All 46 core routes documented."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use 'Authorization: Token ck_...' for customer API key auth"
      },
      "CustomerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Customer API key in format: Token ck_your_api_key_here"
      }
    },
    "schemas": {
      "OkWrapper": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INVALID_API_KEY"
              },
              "message": {
                "type": "string",
                "example": "Invalid or expired customer API key"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "success",
          "error"
        ]
      }
    }
  },
  "paths": {
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "Auth",
          "Admin"
        ],
        "summary": "Admin login (JWT)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkWrapper"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/profile": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get admin profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "admin": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "admin_001"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "example": "admin@yourcompany.com"
                            },
                            "role": {
                              "type": "string",
                              "example": "super_admin"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "example": [
                                "view_all",
                                "manage_customers",
                                "manage_billing",
                                "manage_system"
                              ]
                            },
                            "lastLogin": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update admin profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "newemail@yourcompany.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "admin": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "admin_001"
                            },
                            "email": {
                              "type": "string",
                              "format": "email"
                            },
                            "role": {
                              "type": "string",
                              "example": "super_admin"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "lastLogin": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Admin profile updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/profile": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get admin profile (legacy path)",
        "deprecated": true,
        "description": "Use /api/v1/admin/profile instead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkWrapper"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audit/self": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "View recent admin actions (audit)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entries"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/dev/reset-password": {
      "post": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "DEV ONLY: reset admin password from env",
        "responses": {
          "200": {
            "description": "Reset OK"
          },
          "403": {
            "description": "Forbidden in prod"
          }
        }
      }
    },
    "/api/v1/admin/dev/seed": {
      "post": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "DEV ONLY: seed admin email/password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seeded"
          },
          "403": {
            "description": "Forbidden in prod"
          }
        }
      }
    },
    "/api/v1/admin/debug/verify-token-detailed": {
      "post": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "Verify a plaintext API key against DB",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification details"
          }
        }
      }
    },
    "/api/v1/admin/debug/import-plain-key": {
      "post": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "Import a plaintext key from vault to DB",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plainKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "plainKey"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imported or already exists"
          },
          "400": {
            "description": "Not found in vault"
          }
        }
      }
    },
    "/api/v1/admin/debug/plain-keys": {
      "get": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "List recent masked plaintext keys from file vault",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keys"
          }
        }
      }
    },
    "/api/v1/admin/debug/plain-keys-sqlite": {
      "get": {
        "tags": [
          "Dev/Debug"
        ],
        "summary": "List recent masked keys from sqlite vault",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keys"
          }
        }
      }
    },
...
  }
}